Complex provider for some complex data sources
ComplexDataProvider is a non-abstract class for interactions with some complex data sources like a http-server in a different thread. Notice that OVERRIDING get_content AND result_callback TAKES NO EFFECT.
ComplexDataProvider has a MessageSystem object that is used to transfer data between provider and data source. It is made this way for a proper working of asyncio with threads. MessageSystem is created by complexDataProvider_object.get_message_system.
MessageSystem has 2 main functions:
1. send_to_provider(self, message) - sends a message to provider and returns the message id. You should store this id so you could access the results of processing the message in hooks.
2. retrieve_result(self, message_id) - blocking retrieving of answer for message with message_id.